+++ subpackages/lib/source/odood/lib/deploy/odoo.d
@@ -215,11 +215,12 @@ Project deployOdoo(in DeployConfig config) {
     project.directories.log.setAttributes(octal!750);
 
     // Existing file is required to make fail2ban work.
-    // Thus we just create empty logfile here, odoo will continue to log here
-    if (!project.odoo.logfile.exists) {
-        project.odoo.logfile.writeFile([]);
-        project.odoo.logfile.chown(pw_odoo.pw_uid, pw_odoo.pw_gid);
-        project.odoo.logfile.setAttributes(octal!640);
+    // Thus we just create empty logfile here, odoo will continue to log here.
+    // Skip when logging to stdout (no logfile configured).
+    if (!project.odoo.logfile.isNull && !project.odoo.logfile.get.exists) {
+        project.odoo.logfile.get.writeFile([]);
+        project.odoo.logfile.get.chown(pw_odoo.pw_uid, pw_odoo.pw_gid);
+        project.odoo.logfile.get.setAttributes(octal!640);
     }
 
     // Make Odoo owner of data directory. Do not allow others to access it.