+++ subpackages/lib/source/odood/lib/project/package.d
@@ -2,3 +2,4 @@ module odood.lib.project;
 
 public import odood.lib.project.project;
 public import odood.lib.project.config;
+public import odood.lib.project.info;